refactor: Rename and restructure folders, simplify service
